itmss/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
mock==5.0.1
PyJWT==1.7.1
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
rest-framework-simplejwt==0.0.1
sqlparse==0.4.3
uritemplate==4.1.1
